+++ src/flask/__init__.py
@@ -1,6 +1,3 @@
-from markupsafe import escape
-from markupsafe import Markup
-
 from . import json as json
 from .app import Flask as Flask
 from .app import Request as Request
@@ -67,5 +64,29 @@ def __getattr__(name):
             stacklevel=2,
         )
         return __request_ctx_stack
+
+    if name == "escape":
+        import warnings
+        from markupsafe import escape
+
+        warnings.warn(
+            "'flask.escape' is deprecated and will be removed in Flask 2.4. Import"
+            " 'markupsafe.escape' instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return escape
+
+    if name == "escape":
+        import warnings
+        from markupsafe import Markup
+
+        warnings.warn(
+            "'flask.Markup' is deprecated and will be removed in Flask 2.4. Import"
+            " 'markupsafe.Markup' instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return Markup
 
     raise AttributeError(name)